+++ node/frontend/src/lib/nodes.ts
@@ -1,4 +1,10 @@
-import { CanvasEdge, CanvasNode, FormatReturnType, getNodeContent } from "@coordnet/core";
+import {
+  CanvasEdge,
+  CanvasNode,
+  ExportNode,
+  FormatReturnType,
+  getNodeContent,
+} from "@coordnet/core";
 import { generateJSON, JSONContent } from "@tiptap/core";
 import DOMPurify from "dompurify";
 import { marked } from "marked";
@@ -9,7 +15,7 @@ import * as Y from "yjs";
 import { getNode } from "@/api";
 import { createConnectedCanvas } from "@/components/Canvas/utils";
 import { extensions } from "@/lib/readOnlyEditor";
-import { ExportNode, SpaceNode } from "@/types";
+import { SpaceNode } from "@/types";
 
 import { getCanvas } from "./canvases";
 import { readOnlyEditor } from "./readOnlyEditor";