+++ node/packages/core/src/skills/paperQA.ts
@@ -1,8 +1,14 @@
 import * as Y from "yjs";
 
-import { CanvasNode, PaperQAResponse, PaperQAResponsePair, Task } from "../types";
+import { CanvasNode, PaperQAResponse, PaperQAResponsePair, SingleNode, Task } from "../types";
 import { queryPaperQA } from "./api";
-import { setNodesState, setSkillNodeTitleAndContent } from "./utils";
+import {
+  addToSkillCanvas,
+  findSourceNode,
+  isMultipleResponseNode,
+  setNodesState,
+  setSkillNodeTitleAndContent,
+} from "./utils";
 
 export const formatPaperQAResponse = (data: unknown): PaperQAResponse => {
   if (
@@ -100,22 +106,41 @@ export const executePaperQATask = async (
     let markdown = "";
     try {
       markdown = paperQAResponseToMd(formatPaperQAResponse(response));
-    } catch {
+    } catch (error) {
+      console.error("Error formatting PaperQA response", error);
       try {
         markdown = response[0][1].find((pair: PaperQAResponsePair) => pair[0] === "answer")?.[1];
-      } catch {
+      } catch (error) {
+        console.error("Error parsing PaperQA response", error);
         markdown = JSON.stringify(response, null, 2);
       }
     }
 
-    [task?.outputNode?.id, isLastTask ? outputNode.id : null].forEach(async (id) => {
-      if (id) await setSkillNodeTitleAndContent(skillDoc, id, "PaperQA Response", markdown);
+    const node: SingleNode = { title: "PaperQA Response: " + query, markdown: markdown };
+
+    const sourceNode = findSourceNode(task);
+
+    [task?.outputNode?.id, isLastTask ? outputNode.id : null].forEach(async (canvasId) => {
+      if (!canvasId) return;
+
+      // If it's a multiple response node
+      if (task.outputNode && isMultipleResponseNode(task.outputNode)) {
+        await addToSkillCanvas({ canvasId, document: skillDoc, nodes: [node], sourceNode });
+      } else {
+        // Otherwise just update the node directly
+        await setSkillNodeTitleAndContent(skillDoc, canvasId, node.title, node.markdown);
+      }
     });
 
     // Mark the node as done/inactive
     setNodesState([task.promptNode.id], nodesMap, "inactive");
   } catch (error) {
     console.error("Error executing PaperQA task", error);
-    setNodesState([task.promptNode.id], nodesMap, "inactive");
+    setNodesState(
+      [task.promptNode.id],
+      nodesMap,
+      "error",
+      `PaperQA error: ${(error as Error)?.message}`
+    );
   }
 };