+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+- Ability for Skills to write to Spaces (using External Nodes).
 - Permissions on Skill runs to allow users to share their results.
 
 ## [25.5.2] - 2025-05-19